refactor(signaling): replaced hashed channel IDs with secure random u64 IDs OsRng uses the secure OS random generator mechanism. That mechanism is a CSPRNG (cryptographically secure pseudo-random number generator) Basically its a cryptographic random. - removed DefaultHasher-based channel ID generation - generate channel IDs using secure random u64 values - updated signaling state and packet headers to use u64 channel_id - adjusted client-side channel_id storage and WELCOME parsing - updated packet encode/decode offsets and tests - added tests for secure random channel ID generation

+++ src/signaling/structures.rs
@@ -33,7 +33,7 @@ impl User {
 
 #[derive(Clone, Debug)]
 pub struct Channel {
-    pub channel_id: u32,
+    pub channel_id: u64,
     pub next_peer_id: u32,
     pub users: Vec<User>,
     pub relay: Option<String>,
